+++ src/redischronos.nim
@@ -1,4 +1,4 @@
-const redischronosVersion* = "1.1.1"
+const redischronosVersion* = "1.1.2"
 
 import redischronos/api
 import redischronos/factories